refactor: XMonad help is separated file now; remove: .bash_aliases_extended; refactor: xpropmessage shebang; add: XMonad build checksum source checker; refactor: some aliases moved to zsh;

+++ zsh/INSTALL.sh
@@ -1,12 +1,18 @@
-sudo apt install zsh -y
+#!bin/bash
 
-echo "Install oh my zsh? (1/0)"
-read REPLY
-echo
-if [ $REPLY -eq 1 ]
-then
+echo "Install zsh from apt?"
+read -p "[Yy/Nn]" yn
+if [[ $yn =~ Y|y ]]; then
+    sudo apt install zsh -y
+fi
+
+echo "Install oh my zsh? Use default installation directory."
+read -p "[Yy/Nn]" yn
+if [[ $yn =~ Y|y ]]; then
     sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
 fi
 
-rm ${HOME}/.zshrc
-ln -sv ${HOME}/.dotfiles/zsh/.zshrc $HOME/.zshrc
+if [ -f $HOME/.zshrc ]; then
+    mv -v $HOME/.zshrc $HOME/.zshrc.orig
+fi
+ln -sv $HOME/.dotfiles/zsh/.zshrc $HOME/.zshrc
